rearranged manager.go for virtualizers
